Just day of fun, saving state as sometimes precision of new fft fails

+++ libs/fft/FFTunreal.hpp
@@ -39,7 +39,7 @@ public:
 private:
 
     T* _data_ptr;
-    long _len;
+    uint32_t _len;
 
 };
